day07

+++ src/main/java/leetcode/editor/cn/day03/ValidSudoku.java
@@ -92,7 +92,7 @@ public class ValidSudoku {
                 for (int j = 0; j < 9; j++) {
                     char num = board[i][j];
                     if (num != '.') {
-                        int n = (int) num;
+                        int n = num;
                         int box_index = (i / 3) * 3 + j / 3;
 
                         // keep the current cell value
